허수연 NewBottom 추가

+++ TeamProject/src/main/webapp/board/class/list.jsp
@@ -4,15 +4,13 @@
 <%@ page language="java" contentType="text/html; charset=UTF-8"
     pageEncoding="UTF-8"%>
     <!DOCTYPE html>
-<html>
 <head>
+  <title>Shop</title> 
+  <meta charset="utf-8">
   <link rel="stylesheet" href="shopcss.css">
   <link href="css/NewTop.css" rel="stylesheet" type="text/css">
+  <link href="css/NewBottom.css" rel="stylesheet" type="text/css">
   <link href="css/product.css" rel="stylesheet" type="text/css">
-  <link href="https://fonts.googleapis.com/earlyaccess/notosanskr.css" rel="stylesheet">
-  <title>Shop</title> 
-  </style>
-  <meta charset="utf-8">
   <link rel="stylesheet" href="shopcss.css">
   <link href="https://fonts.googleapis.com/earlyaccess/notosanskr.css" rel="stylesheet">
 </head>
@@ -24,12 +22,10 @@
 List<ClassBoardDTO> boardList =(List<ClassBoardDTO>)request.getAttribute("boardList");
 PageDTO pageDTO = (PageDTO)request.getAttribute("pageDTO");
 String memberId = (String)session.getAttribute("memberId");
+String memberType = (String)session.getAttribute("memberType");
 %>
-
     <div class="products">
         <h3>클래스 목록</h3>
-        
-
         <div class="product-list">        
         <%
 for(int i=0; i<boardList.size(); i++){
@@ -56,14 +52,16 @@ for(int i=0; i<boardList.size(); i++){
 <%
 }
 %>
-
-   <input type="button" value="클래스 등록" class="btn" 
+<% if(memberId != null){ %>
+    <% if(memberType.equals("host")){ %>
+  <input type="button" value="클래스 등록" class="btn" 
        					onclick="location.href='classWrite.cbo'">
   <input type="button" value="클래스 수정" class="btn" 
        					onclick="location.href='classUpdate.cbo'">
   <input type="button" value="클래스 삭제" class="btn" 
        					onclick="location.href='classDelete.cbo'">
-       					
+     <% } %>
+ <% } %>
  <div id="page_control">
 <%
 // 시작페이지 1페이지 prev 없음 
@@ -89,13 +87,11 @@ if(pageDTO.getEndPage() < pageDTO.getPageCount()){
 	<%
 }
 %>
-
 </div> 
     </div>
     
 <script type="text/javascript"></script>
 <script src ="js/shop__buttonContainer.js"></script>
-
     
     <!-- 푸터들어가는 곳 -->
 <jsp:include page="../../inc/bottom.jsp"></jsp:include>
